Make `packages` file obsolete
The git URL for the repository of a package follows a standard naming
convention (obviously). Instead of keeping a file that maps custom names
towards git URLs we just supply the package name as named in AUR and do
the rest from there. No need to store what packages are installed
because the script already assumes that all directories are AUR packages
that the user wants to keep updated.